Version 1 - Display date is implemented
A lot of time lost on ListView when I found out it lacked a few essential things to make my vision true. Therefore had to rebuild the display using TableView. A few more fields have been added to the list: food, tasks, and empty field for delete and edit buttons.